CIF-2120: Refactor storefront context packages
