+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/product/Product.java
@@ -19,7 +19,7 @@ import java.util.List;
 import com.adobe.cq.commerce.core.components.models.common.Price;
 import com.adobe.cq.commerce.core.components.models.page.PageMetadata;
 import com.adobe.cq.commerce.core.components.models.retriever.AbstractProductRetriever;
-import com.adobe.cq.commerce.core.components.schemas.ProductSchema;
+import com.adobe.cq.commerce.core.components.storefrontcontext.ProductStorefrontContext;
 import com.adobe.cq.wcm.core.components.models.Component;
 
 /**
@@ -89,5 +89,5 @@ public interface Product extends Component, PageMetadata {
 
     AbstractProductRetriever getProductRetriever();
 
-    ProductSchema getProductData();
+    ProductStorefrontContext getProductData();
 }
